Inline StaffAttendanceService into StaffAttendanceController
